global: fix std::move misuse and other clang warnings

+++ core/common/src/image.cpp
@@ -13,7 +13,7 @@ namespace bb
     {
       assert((copy.width != 0) && (copy.height != 0) && (copy.depth != 0));
 
-      size_t dataSize = copy.width * copy.height * copy.depth;
+      size_t dataSize = static_cast<size_t>(copy.width * copy.height * copy.depth);
       this->data.reset(new uint8_t[dataSize]);
 
       memcpy(this->data.get(), copy.data.get(), dataSize);
@@ -38,7 +38,7 @@ namespace bb
     { // has data to copy
       assert((copy.width != 0) && (copy.height != 0) && (copy.depth != 0));
 
-      size_t dataSize = copy.width * copy.height * copy.depth;
+      size_t dataSize = static_cast<size_t>(copy.width * copy.height * copy.depth);
       this->data.reset(new uint8_t[dataSize]);
 
       memcpy(this->data.get(), copy.data.get(), dataSize);
